+++ collector/src/test/java/com/bitmovin/analytics/error/ErrorReportingLimiterTest.kt
@@ -1,5 +1,8 @@
 package com.bitmovin.analytics.error
 
+import com.bitmovin.analytics.api.error.ErrorSeverity
+import com.bitmovin.analytics.dtos.ErrorCode
+import com.bitmovin.analytics.dtos.ErrorData
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 
@@ -9,10 +12,26 @@ class ErrorReportingLimiterTest {
         val identicalErrorReportingLimiter = IdenticalErrorReportingLimiter()
 
         for (i in 1..5) {
-            assertThat(identicalErrorReportingLimiter.shouldReportError(10)).isTrue
+            assertThat(
+                identicalErrorReportingLimiter.shouldReportError(
+                    ErrorCode(
+                        10,
+                        "testMessage",
+                        ErrorData(),
+                    ),
+                ),
+            ).isTrue
         }
 
-        assertThat(identicalErrorReportingLimiter.shouldReportError(10)).isFalse
+        assertThat(
+            identicalErrorReportingLimiter.shouldReportError(
+                ErrorCode(
+                    10,
+                    "testMessage",
+                    ErrorData(),
+                ),
+            ),
+        ).isFalse
     }
 
     @Test
@@ -20,10 +39,26 @@ class ErrorReportingLimiterTest {
         val identicalErrorReportingLimiter = IdenticalErrorReportingLimiter()
 
         for (i in 1..5) {
-            assertThat(identicalErrorReportingLimiter.shouldReportError(10)).isTrue
+            assertThat(
+                identicalErrorReportingLimiter.shouldReportError(
+                    ErrorCode(
+                        10,
+                        "testMessage1",
+                        ErrorData(),
+                    ),
+                ),
+            ).isTrue
         }
 
-        assertThat(identicalErrorReportingLimiter.shouldReportError(5)).isTrue
+        assertThat(
+            identicalErrorReportingLimiter.shouldReportError(
+                ErrorCode(
+                    5,
+                    "testMessage2",
+                    ErrorData(),
+                ),
+            ),
+        ).isTrue
     }
 
     @Test
@@ -32,8 +67,53 @@ class ErrorReportingLimiterTest {
 
         // two alternating errors (which is not blocked)
         for (i in 1..10) {
-            assertThat(identicalErrorReportingLimiter.shouldReportError(10)).isTrue
-            assertThat(identicalErrorReportingLimiter.shouldReportError(5)).isTrue
+            assertThat(
+                identicalErrorReportingLimiter.shouldReportError(
+                    ErrorCode(
+                        10,
+                        "testMessage1",
+                        ErrorData(),
+                    ),
+                ),
+            ).isTrue
+            assertThat(
+                identicalErrorReportingLimiter.shouldReportError(
+                    ErrorCode(
+                        5,
+                        "testMessage2",
+                        ErrorData(),
+                    ),
+                ),
+            ).isTrue
+        }
+    }
+
+    @Test
+    fun testErrorReportingLimiter_ShouldReportAllErrorsWithAlternatingSeverity() {
+        val identicalErrorReportingLimiter = IdenticalErrorReportingLimiter()
+
+        // two alternating errors (which is not blocked)
+        for (i in 1..10) {
+            assertThat(
+                identicalErrorReportingLimiter.shouldReportError(
+                    ErrorCode(
+                        10,
+                        "testMessage1",
+                        ErrorData(),
+                        errorSeverity = ErrorSeverity.CRITICAL,
+                    ),
+                ),
+            ).isTrue
+            assertThat(
+                identicalErrorReportingLimiter.shouldReportError(
+                    ErrorCode(
+                        10,
+                        "testMessage1",
+                        ErrorData(),
+                        errorSeverity = ErrorSeverity.INFO,
+                    ),
+                ),
+            ).isTrue
         }
     }
 
@@ -42,11 +122,27 @@ class ErrorReportingLimiterTest {
         val identicalErrorReportingLimiter = IdenticalErrorReportingLimiter()
 
         for (i in 1..5) {
-            assertThat(identicalErrorReportingLimiter.shouldReportError(10)).isTrue
+            assertThat(
+                identicalErrorReportingLimiter.shouldReportError(
+                    ErrorCode(
+                        10,
+                        "testMessage",
+                        ErrorData(),
+                    ),
+                ),
+            ).isTrue
         }
 
         identicalErrorReportingLimiter.reset()
 
-        assertThat(identicalErrorReportingLimiter.shouldReportError(10)).isTrue
+        assertThat(
+            identicalErrorReportingLimiter.shouldReportError(
+                ErrorCode(
+                    10,
+                    "testMessage",
+                    ErrorData(),
+                ),
+            ),
+        ).isTrue
     }
 }